pyproject.toml
[tool.poetry]
name = "awstools"
version = "0.1.0"
description = ""
authors = ["Daniel Souza <daniel.souza@oxygea.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
boto3 = "^1.35.53"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
